requirements.txt
discord.py==2.4.0
discord-py-interactions==4.4.0
gspread==6.1.4
oauth2client==4.1.3
python-dotenv==1.0.1
rapidfuzz==3.10.1
numpy==2.2.1
pytz==2024.2